requirements.txt
actionlib==1.14.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
angles==1.9.13
anyio==3.7.1
apptools==5.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.2.0
autolab-core==1.1.1
autolab-perception==1.0.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
bondpy==1.8.6
bop-toolkit-lib==1.0
braceexpand==0.1.7
cached-property==1.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
colorlog==6.8.2
comm==0.1.4
configobj==5.0.8
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cvxopt==1.3.2
cycler==0.11.0
Cython==0.29.24
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.3.7
dynamic-reconfigure==1.7.3
easy-dexnet==0.1.0
entrypoints==0.4
envisage==7.0.3
exceptiongroup==1.2.1
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
fonttools==4.38.0
fqdn==1.5.1
freetype-py==2.4.0
future==1.0.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
glumpy==1.1.0
grasp-nms==1.0.2
graspnetAPI==1.2.11
h5py==3.8.0
hsluv==5.0.4
idna==3.7
image-geometry==1.16.2
imageio==2.5.0
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
interactive-markers==1.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.16
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter-events==0.6.3
jupyter_packaging==0.12.3
jupyter-server==1.24.0
jupyter_server_fileid==0.9.2
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.11
kiwisolver==1.3.1
laser_geometry==1.6.7
MarkupSafe==2.1.5
matplotlib==2.2.4
matplotlib-inline==0.1.6
mayavi==4.7.4
message-filters==1.16.0
mistune==3.0.2
moveit-core==1.1.13
moveit-ros-planning-interface==1.1.13
multiprocess==0.70.15
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
open3d==0.13.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
plyfile==0.9
prometheus-client==0.17.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pybullet==3.2.6
pycocotools==2.0
pycparser==2.21
pyface==8.0.0
pyglet==1.4.10
Pygments==2.17.2
pyhull==2015.2.1
PyOpenGL==3.1.0
pyparsing==3.1.2
pypng==0.0.19
pyrender==0.1.45
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-qt-binding==0.4.4
pytz==2024.1
PyWavefront==1.3.3
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==26.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.31.0
resource_retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt-console==0.4.12
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt-image-view==0.4.17
rqt_launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt_robot_steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt-tf-tree==0.6.4
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rviz==1.14.20
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.5.1
Send2Trash==1.8.3
sensor-msgs==1.13.1
setproctitle==1.3.3
setuptools==65.6.3
shapely==2.0.6
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
soupsieve==2.4.1
srdfdom==0.6.4
terminado==0.17.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.5
topic-tools==1.16.0
torch==1.10.1
torchvision==0.11.2
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
traits==6.4.3
traitsui==8.0.0
transforms3d==0.3.1
triangle==20230923
trimesh==4.4.1
typing_extensions==4.7.1
urdfdom-py==0.4.6
uri-template==1.3.0
urllib3==2.0.7
vispy==0.13.0
vtk==8.1.2
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.6.1
wheel==0.38.4
widgetsnbextension==4.0.11
xacro==1.14.18
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.15.0
